Removing global stuff for now
